Move sourcefile construction to the manifest update function
Instead of having the tree iterator return either paths or constructed
sourcefiles, just have it always return paths, and construct the
sourcefiles in the caller. This has the advantage that we can better
handle some error conditions that may be caused by having an invalid
cache. It is also somewhat clearer since we don't end up with an
iterator that can return quite different types depending on whether it
thinks the file was updated or not.