Implement the media element 'show poster flag' in accordance with spec
Previously, HTMLMediaElement had a 'DisplayMode' enum which was one of
{ 'kUnknown', 'kPoster', 'kVideo' }. This didn't mesh well with the
spec, which only mentions the 'show poster flag' as being true or false.

Additionally, the points where this mode was changed were inconsistent
with the spec, and resulted in us showing video contents when we should
be showing the poster, and makes implementing other features dependent
on the state of the 'show poster flag' difficult, such as the cue
timeline.

This CL removes 'DisplayMode' from HTMLMediaElement (though reintroduces
it in a limited form in LayoutVideo), and implements the
'show-poster-flag' in accordance with the spec. This makes the behavior
of poster rendering (as detailed in the attached bug) consistent with
Firefox and Safari.

Bug: 1007241
Change-Id: Iaaf3a10581357443b7889578360852e0a674e68c