Extract a list of valid URLs from the resources attribute value
We need a list of valid URLs from <link> element's resources attribute
value so that we can tell whether subresources should be loaded from
the WebBundle or not.

This CL does:
- Change <link> element's resources= attribute to DOMTokenList
- Make HTMLLinkElement have |valid_resource_urls_|, which is a list of
  valid urls and will be used later. |valid_resource_urls_| should be
  updated whenever resources attribute changes.

Note that changing the type of resources attribute to DOMTokenList is not
mandatory here. That's just for web developer ergonomics.

Change-Id: Ic10abab4bf8e07b245fea7e39becc5ab1aa2c82a