Split up canvas test yaml files
canvas2d tests in the wpt directory are stored in gigantic (>10,000
line) files. This CL attempts to split them up by theme. The script
for splitting them is here:
https://github.com/mysteryDate/wpt-helper-scripts/blob/master/split_yaml.py

The biggest files are still quite large, but it's a start. Also I think
there's some redundancy between the offscreen and element versions that
could be addressed later. Running build.sh now doesn't change any files,
which is to be expected.

Bug: 1093484, 1071217
Change-Id: Id8fb79df3134e5f4e504c1c57e0da679f09afc64